Migrate discal to discord.calendar
