+++ .coveragerc
@@ -216,8 +216,7 @@ omit =
     homeassistant/components/discogs/sensor.py
     homeassistant/components/discord/__init__.py
     homeassistant/components/discord/notify.py
-    homeassistant/components/discal/__init__.py
-    homeassistant/components/discal/calendar.py
+    homeassistant/components/discord/calendar.py
     homeassistant/components/dlib_face_detect/image_processing.py
     homeassistant/components/dlib_face_identify/image_processing.py
     homeassistant/components/dlink/switch.py
