+++ CODEOWNERS
@@ -244,11 +244,9 @@ build.json @home-assistant/supervisor
 /homeassistant/components/diagnostics/ @home-assistant/core
 /tests/components/diagnostics/ @home-assistant/core
 /homeassistant/components/digital_ocean/ @fabaff
-/homeassistant/components/discal/ @GrandMoff100
-/tests/components/discal/ @GrandMoff100
 /homeassistant/components/discogs/ @thibmaek
-/homeassistant/components/discord/ @tkdrob
-/tests/components/discord/ @tkdrob
+/homeassistant/components/discord/ @tkdrob @GrandMoff100
+/tests/components/discord/ @tkdrob @GrandMoff100
 /homeassistant/components/discovery/ @home-assistant/core
 /tests/components/discovery/ @home-assistant/core
 /homeassistant/components/dlna_dmr/ @StevenLooman @chishm
