--- homeassistant/components/discal/const.py
@@ -1,3 +1,0 @@
-"""Constants for the Discord Calendar integration."""
-
-DOMAIN = "discal"
