--- homeassistant/components/discal/manifest.json
@@ -1,13 +1,0 @@
-{
-  "domain": "discal",
-  "name": "Discord Calendar",
-  "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/discal",
-  "requirements": [],
-  "ssdp": [],
-  "zeroconf": [],
-  "homekit": {},
-  "dependencies": ["discord", "persistent_notification"],
-  "codeowners": ["@GrandMoff100"],
-  "iot_class": "cloud_polling"
-}
