--- homeassistant/components/discal/strings.json
@@ -1,22 +1,0 @@
-{
-  "config": {
-    "step": {
-      "user": {
-        "data": {
-          "token": "[%key:common::config_flow::data::api_token%]",
-          "guild_ids": "Guild IDs"
-        }
-      }
-    },
-    "error": {
-      "connection_error": "[%key:common::config_flow::error::cannot_connect%]",
-      "invalid_token": "[%key:common::config_flow::error::invalid_access_token%]",
-      "unknown": "[%key:common::config_flow::error::unknown%]",
-      "server_not_found": "One or more Discord servers could not be found.",
-      "malformed_guilds": "Malformed Guild IDs field."
-    },
-    "abort": {
-      "already_configured": "[%key:common::config_flow::abort::already_configured_device%]"
-    }
-  }
-}
