--- homeassistant/components/discal/translations/en.json
@@ -1,22 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Device is already configured"
-        },
-        "error": {
-            "connection_error": "Failed to connect",
-            "invalid_token": "Invalid access token",
-            "malformed_guilds": "Malformed Guild IDs field.",
-            "server_not_found": "One or more Discord servers could not be found.",
-            "unknown": "Unexpected error"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "guild_ids": "Guild IDs",
-                    "token": "API Token"
-                }
-            }
-        }
-    }
-}
