+++ homeassistant/components/discord/__init__.py
@@ -10,23 +10,22 @@ from homeassistant.helpers import discovery
 
 from .const import DOMAIN
 
-PLATFORMS = [Platform.NOTIFY]
+PLATFORMS = [Platform.NOTIFY, Platform.CALENDAR]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Discord from a config entry."""
     nextcord.VoiceClient.warn_nacl = False
-    discord_bot = nextcord.Client()
+    hass.data.setdefault(DOMAIN, {}).setdefault(
+        entry.entry_id,
+        dict(entry.data),
+    )["client"] = discord_bot = nextcord.Client()
     try:
         await discord_bot.login(entry.data[CONF_API_TOKEN])
     except nextcord.LoginFailure as ex:
         raise ConfigEntryAuthFailed("Invalid token given") from ex
     except (ClientConnectorError, nextcord.HTTPException, nextcord.NotFound) as ex:
         raise ConfigEntryNotReady("Failed to connect") from ex
-    finally:
-        await discord_bot.close()
-
-    hass.data.setdefault(DOMAIN, {})[entry.entry_id] = entry.data
 
     hass.async_create_task(
         discovery.async_load_platform(
@@ -37,5 +36,14 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
             hass.data[DOMAIN],
         )
     )
+    hass.config_entries.async_setup_platforms(entry, [Platform.CALENDAR])
 
     return True
+
+
+async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Close the discord client connection and pop the shared data."""
+    if unload_ok := await hass.config_entries.async_unload_platforms(entry, PLATFORMS):
+        await hass.data[DOMAIN][entry.entry_id]["client"].close()
+        hass.data[DOMAIN].pop(entry.entry_id)
+    return unload_ok
