+++ homeassistant/components/discord/manifest.json
@@ -4,7 +4,7 @@
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/discord",
   "requirements": ["nextcord==2.0.0a8"],
-  "codeowners": ["@tkdrob"],
+  "codeowners": ["@tkdrob", "@GrandMoff100"],
   "iot_class": "cloud_push",
   "loggers": ["discord"]
 }
