+++ mypy.ini
@@ -610,17 +610,6 @@ no_implicit_optional = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.discal.*]
-check_untyped_defs = true
-disallow_incomplete_defs = true
-disallow_subclassing_any = true
-disallow_untyped_calls = true
-disallow_untyped_decorators = true
-disallow_untyped_defs = true
-no_implicit_optional = true
-warn_return_any = true
-warn_unreachable = true
-
 [mypy-homeassistant.components.dlna_dmr.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
