--- tests/components/discal/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Discord Calendar integration."""
